Continue refactor of publishers
